# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: My REST API
  description: Some custom description of API.
  contact:
    name: Artem Yushkevich
    url: ''
    email: yushkevich@mail.ru
  license:
    name: License of API
    url: http://
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Dev server
security:
  - Bearer-Authentication: [ ]
paths:
  /api/v1/user/update:
    put:
      tags:
        - User controller
      operationId: updateUserInfo
      description: Allows you to update user info such as name, surname, city, street or building.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/catalog/product/{productId}/reviews/create:

    post:
      description: Allows the user to create new review on the product
      tags:
        - Product controller
      operationId: createNewReview
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/user/reviews:
    get:
      description: Allows you to retrieve user reviews
      tags:
        - User controller
      operationId: getUserReviews
      parameters:
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/GetUserReviewsDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserReviewsDTO'
  /api/v1/user/orders:
    get:
      description: Allow you to retrieve user orders
      tags:
        - User controller
      operationId: getUserOrders
      parameters:
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/GetUserOrderPageDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrdersDTO'
  /api/v1/user/orders/{orderId}/items:
    get:
      description: Allows you to retrieve all order items in the order by order ID
      tags:
        - User controller
      operationId: getUserOrderItems
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/GetOrderItemsDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemsDTO'
...
  /api/v1/catalog/product/{productId}/reviews:
    get:
      description: Allows you to retrieve all reviews for a product by their ID
      tags:
        - Product controller
      operationId: getProductReviewsById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/GetReviewsDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReviewsDTO'
  /api/v1/catalog/product/{productId}/reviews/user_review:
    get:
      description: Allows you to view user reviews for a product, if available
      tags:
        - Product controller
      operationId: getUserReviewOnProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReviewDTO'
  /api/v1/catalog/product/{productId}/:
    get:
      description: Allows you to view the product page with additional information
      tags:
        - Product controller
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
...
  /api/v1/catalog/product/{productId}/reviews/delete:
    delete:
      description: Allows you to delete your own review
      tags:
        - Product controller
      operationId: deleteReview
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/admin/get/order/{orderId}:
    delete:
      description: Allows you to get the items from the order
      tags:
        - Admin controller
      operationId: getOrderItems
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderItemsDTO'
...
components:
  schemas:
    UpdateUserDTO:
      description: New user data
      required:
        - building
        - city
        - name
        - street
        - surname
      type: object
      properties:
        name:
          type: string
          default: "New name"
          description: New user name
        surname:
          type: string
          default: "New surname"
          description: New user surname
        city:
          default: "New city"
          description: New user city
          type: string
        street:
          default: "New street"
          description: New user street
          type: string
        building:
          default: 123
          description: New user building
          maximum: 1000
          minimum: 1
          type: integer
          format: int32
...
    CreateReviewDTO:
      description: Review data for creating a new review
      required:
        - description
        - rating
        - userId
      type: object
      properties:
        description:
          default: "Description"
          description: Review details
          type: string
        rating:
          default: 5
          description: Review rating
          maximum: 10
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
          format: double
        productId:
          default: 1
          description: Target product ID
          type: integer
          format: int64
        userId:
          default: 1
          description: User ID who created the review
          minimum: 1
          type: integer
          format: int64
...
    GetUserReviewsDTO:
      description: Required data for user feedback on a page-by-page basis
      required:
        - countPerPage
        - pageNum
      type: object
      properties:
        pageNum:
          default: 1
          description: Num of the page
          minimum: 1
          type: integer
          format: int32
        countPerPage:
          default: 10
          description: Number of elements on the page
          minimum: 10
          type: integer
          format: int32
    UserReviewInfoDTO:
      description: User review information
      required:
        - description
        - imageLink
        - productId
        - rating
      type: object
      properties:
        description:
          description: Review details
          type: string
        rating:
          description: Review rating
          maximum: 10
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
          format: double
        productId:
          description: Product ID
          minimum: 1
          type: integer
          format: int64
        imageLink:
          description: Product image link
          type: string
        name:
          description: Product name
          pattern: ^[A-zА-я0-9]{1,40}$
          type: string
    UserReviewsDTO:
      description: List of user reviews with the total number of reviews of this user in the database
      required:
        - countOf
        - reviews
      type: object
      properties:
        reviews:
          description: List of user reviews
          type: array
          items:
            $ref: '#/components/schemas/UserReviewInfoDTO'
        countOf:
          description: Total number of user reviews
          minimum: 0
          type: integer
          format: int64
    GetUserOrderPageDTO:
      description: Required data for getting the user orders on a page-by-page basis
      required:
        - countPerPage
        - pageNum
      type: object
      properties:
        pageNum:
          description: Num of the page
          default: 1
          minimum: 1
          type: integer
          format: int32
        countPerPage:
          default: 10
          description: Number of elements on the page
          minimum: 10
          type: integer
          format: int32
    OrderDTO:
      description: Order information
      required:
        - countOfItems
        - date
        - id
        - orderStatus
      type: object
      properties:
        id:
          description: Order ID
          minimum: 1
          type: integer
          format: int64
        date:
          description: Date of order creation
          type: string
          format: date-time
        orderStatus:
          description: Order status
          type: string
          enum:
            - PENDING
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELLED
        countOfItems:
          description: Count of different products in an order
          minimum: 0
          type: integer
          format: int64
    OrdersDTO:
      description: Contains a list of orders and the total number of orders in the database
      required:
        - count
        - orderList
      type: object
      properties:
        orderList:
          description: Order list
          type: array
          items:
            $ref: '#/components/schemas/OrderDTO'
        count:
          description: Total number of orders in the database
          minimum: 0
          type: integer
          format: int64
    GetOrderItemsDTO:
      description: Required data for receipt of all items which order contains
      required:
        - countPerPage
        - orderId
        - pageNum
      type: object
      properties:
        pageNum:
          default: 1
          description: Page number
          minimum: 1
          type: integer
          format: int32
        countPerPage:
          default: 10
          description: Number of elements on the page
          minimum: 10
          type: integer
          format: int32
        orderId:
          default: 1
          description: Target Order ID
          minimum: 1
          type: integer
          format: int64
    OrderItemProductDTO:
      description: Information about the product that is included in the order
      required:
        - count
        - imageLink
        - name
        - price
        - productId
      type: object
      properties:
        productId:
          description: Target product ID
          minimum: 1
          type: integer
          format: int64
        count:
          description: Product count
          minimum: 0
          type: integer
          format: int64
        price:
          description: Product price
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          format: double
        name:
          description: Product name
          pattern: ^[A-zА-я0-9]{1,40}$
          type: string
        imageLink:
          description: Product image link
          type: string
    OrderItemsDTO:
      required:
        - orderItemProductDTOList
        - totalCountOfItems
      type: object
      properties:
        orderItemProductDTOList:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemProductDTO'
        totalCountOfItems:
          minimum: 0
          type: integer
          format: int64
...
    ProductDTO:
      description: Product data
      required:
        - categoryId
        - description
        - id
        - imageLink
        - name
        - price
      type: object
      properties:
        id:
          description: Product ID
          minimum: 1
          type: integer
          format: int64
        categoryId:
          description: Category ID
          minimum: 1
          type: integer
          format: int64
        name:
          description: Product name
          pattern: ^[A-zА-я0-9]{1,40}$
          type: string
        description:
          description: Product description
          type: string
        price:
          description: Product price
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          format: double
        imageLink:
          description: Product image link
          type: string
        rating:
          description: Product rating
          type: number
          format: double
...
    GetReviewsDTO:
      required:
        - countPerPage
        - pageNum
      type: object
      properties:
        pageNum:
          minimum: 1
          type: integer
          format: int32
        countPerPage:
          minimum: 10
          type: integer
          format: int32
        productId:
          type: integer
          format: int64
    ReviewDTO:
      required:
        - description
        - name
        - rating
        - surname
        - userId
      type: object
      properties:
        userId:
          minimum: 1
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        description:
          type: string
        rating:
          maximum: 10
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
          format: double
    ReviewsDTO:
      required:
        - countOfReviews
        - reviewDTOList
      type: object
      properties:
        reviewDTOList:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDTO'
        countOfReviews:
          minimum: 0
          type: integer
          format: int64
...
  securitySchemes:
    Bearer-Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT